Feat/improve UI (#174)
* feat(signup): add sign in page link

* feat(upload): improve ui

* ui(header): add logout button

* feat(login): add redirection for logged user
