+++ frontend/app/upload/components/Crawler/helpers/isValidUrl.ts
@@ -1,0 +1,8 @@
+export const isValidUrl = (string: string) => {
+  try {
+    new URL(string);
+    return true;
+  } catch (_) {
+    return false;
+  }
+};
